// src/generated-api/src/models/DomainAnalyticsPublisherResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Affiliate Backend API
 * API Server for Affiliate Backend Application
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DomainAnalyticsPublisherResponsePublisher } from './DomainAnalyticsPublisherResponsePublisher';
import {
    DomainAnalyticsPublisherResponsePublisherFromJSON,
    DomainAnalyticsPublisherResponsePublisherFromJSONTyped,
    DomainAnalyticsPublisherResponsePublisherToJSON,
    DomainAnalyticsPublisherResponsePublisherToJSONTyped,
} from './DomainAnalyticsPublisherResponsePublisher';

/**
 * 
 * @export
 * @interface DomainAnalyticsPublisherResponse
 */
export interface DomainAnalyticsPublisherResponse {
    /**
     * 
     * @type {DomainAnalyticsPublisherResponsePublisher}
     * @memberof DomainAnalyticsPublisherResponse
     */
    publisher?: DomainAnalyticsPublisherResponsePublisher;
}

/**
 * Check if a given object implements the DomainAnalyticsPublisherResponse interface.
 */
export function instanceOfDomainAnalyticsPublisherResponse(value: object): value is DomainAnalyticsPublisherResponse {
    return true;
}

export function DomainAnalyticsPublisherResponseFromJSON(json: any): DomainAnalyticsPublisherResponse {
    return DomainAnalyticsPublisherResponseFromJSONTyped(json, false);
}

export function DomainAnalyticsPublisherResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): DomainAnalyticsPublisherResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'publisher': json['publisher'] == null ? undefined : DomainAnalyticsPublisherResponsePublisherFromJSON(json['publisher']),
    };
}

export function DomainAnalyticsPublisherResponseToJSON(json: any): DomainAnalyticsPublisherResponse {
    return DomainAnalyticsPublisherResponseToJSONTyped(json, false);
}

export function DomainAnalyticsPublisherResponseToJSONTyped(value?: DomainAnalyticsPublisherResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'publisher': DomainAnalyticsPublisherResponsePublisherToJSON(value['publisher']),
    };
}